+++ src/security/digest/sha1/sha1MessageDigest.cpp
@@ -108,10 +108,10 @@ void sha1MessageDigest::update(const byte_t* buffer, const unsigned long len)
 
 	j = (m_count[0] >> 3) & 63;
 
-	if ((m_count[0] += len << 3) < (len << 3))
+	if ((m_count[0] += static_cast <unsigned int>(len << 3)) < static_cast <unsigned int>(len << 3))
 		m_count[1]++;
 
-	m_count[1] += (len >> 29);
+	m_count[1] += static_cast <unsigned int>(len >> 29);
 
 	if ((j + len) > 63)
 	{
